requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.10.0
Pillow==9.0.1
pandas
fpdf
inquirer